refactor(storage): align AuthorizationSessionRepository.DeleteExpired return type to int
All other repository DeleteExpired methods return (int, error).
AuthorizationSessionRepository was the only outlier using int64.
Align for consistency.

+++ internal/adapters/http/handlers/consent/grants_handler_test.go
@@ -1108,7 +1108,7 @@ func (m *mockAuthSessionRepo) Consume(ctx context.Context, sessionID string) err
 	return nil
 }
 
-func (m *mockAuthSessionRepo) DeleteExpired(_ context.Context) (int64, error) {
+func (m *mockAuthSessionRepo) DeleteExpired(_ context.Context) (int, error) {
 	return 0, nil
 }
